feat(electron): add Electron entrypoint and make OpenFeature self-registering
Adds Tracer.registerFeature() / NoopProxy.registerNoop() so subsystems
can opt into the tracer without being hardcoded in proxy.js. OpenFeature
uses this via a new register.js that self-registers on require.

The Electron entrypoint (index.electron.js) simply omits that require,
so OpenFeature and its native dep (@datadog/openfeature-node-server) are
absent from the Electron build with no conditional logic in the core.
index.js delegates to index.electron.js after calling register.js,
keeping both entrypoints in sync with zero duplication.
