+++ index.electron.js
@@ -1,0 +1,3 @@
+'use strict'
+
+module.exports = require('./packages/dd-trace/index.electron')
