+++ package.json
@@ -144,6 +144,7 @@
     "webpack.js",
     "ext/**/*",
     "index.d.ts",
+    "index.electron.js",
     "index.js",
     "init.js",
     "initialize.mjs",
@@ -153,6 +154,7 @@
     "LICENSE.BSD3",
     "loader-hook.mjs",
     "packages/*/index.js",
+    "packages/*/index.electron.js",
     "packages/*/lib/**/*",
     "packages/*/src/**/*",
     "packages/datadog-instrumentations/orchestrion.yml",
