+++ packages/dd-trace/src/noop/proxy.js
@@ -2,7 +2,6 @@
 
 const NoopAppsecSdk = require('../appsec/sdk/noop')
 const NoopLLMObsSDK = require('../llmobs/noop')
-const NoopFlaggingProvider = require('../openfeature/noop')
 const NoopAIGuardSDK = require('../aiguard/noop')
 const NoopDogStatsDClient = require('./dogstatsd')
 const NoopTracer = require('./tracer')
@@ -11,7 +10,6 @@ const noop = new NoopTracer()
 const noopAppsec = new NoopAppsecSdk()
 const noopDogStatsDClient = new NoopDogStatsDClient()
 const noopLLMObs = new NoopLLMObsSDK(noop)
-const noopOpenFeatureProvider = new NoopFlaggingProvider()
 const noopAIGuard = new NoopAIGuardSDK()
 const noopProfiling = {
   setCustomLabelKeys () {},
@@ -20,13 +18,23 @@ const noopProfiling = {
 
 /** @type {import('../../src/index')} Proxy */
 class NoopProxy {
+  static _noopRegistry = {}
+
+  /**
+   * @param {string} name
+   * @param {object} value
+   */
+  static registerNoop (name, value) {
+    NoopProxy._noopRegistry[name] = value
+  }
+
   constructor () {
     this._tracer = noop
     this.appsec = noopAppsec
     this.dogstatsd = noopDogStatsDClient
     this.llmobs = noopLLMObs
-    this.openfeature = noopOpenFeatureProvider
     this.aiguard = noopAIGuard
+    Object.assign(this, NoopProxy._noopRegistry)
     this.setBaggageItem = (key, value) => {}
     this.getBaggageItem = (key) => {}
     this.getAllBaggageItems = () => {}
