+++ packages/dd-trace/src/proxy.js
@@ -69,6 +69,15 @@ function defineLazily (obj, property, getClass, ...args) {
 }
 
 class Tracer extends NoopProxy {
+  static _featureRegistry = []
+
+  /**
+   * @param {{ name: string, factory: () => object, remoteConfig?: Function, enable?: Function }} feature
+   */
+  static registerFeature (feature) {
+    Tracer._featureRegistry.push(feature)
+  }
+
   constructor () {
     super()
 
@@ -91,7 +100,10 @@ class Tracer extends NoopProxy {
       iast: new LazyModule(() => require('./appsec/iast')),
       llmobs: new LazyModule(() => require('./llmobs')),
       rewriter: new LazyModule(() => require('./appsec/iast/taint-tracking/rewriter')),
-      openfeature: new LazyModule(() => require('./openfeature')),
+    }
+
+    for (const feature of Tracer._featureRegistry) {
+      this._modules[feature.name] = new LazyModule(feature.factory)
     }
   }
 
@@ -176,8 +188,9 @@ class Tracer extends NoopProxy {
           DynamicInstrumentation.start(config, rc)
         }
 
-        const openfeatureRemoteConfig = require('./openfeature/remote_config')
-        openfeatureRemoteConfig.enable(rc, config, () => this.openfeature)
+        for (const feature of Tracer._featureRegistry) {
+          feature.remoteConfig?.(rc, config, this)
+        }
       }
 
       if (config.profiling.DD_PROFILING_ENABLED === 'true') {
@@ -292,9 +305,8 @@ class Tracer extends NoopProxy {
         }
         this._tracingInitialized = true
       }
-      if (config.experimental.flaggingProvider.enabled) {
-        this._modules.openfeature.enable(config)
-        lazyProxy(this, 'openfeature', () => require('./openfeature/flagging_provider'), this._tracer, config)
+      for (const feature of Tracer._featureRegistry) {
+        feature.enable?.(config, this._tracer, this, lazyProxy)
       }
       if (config.iast.enabled) {
         this._modules.iast.enable(config, this._tracer)
@@ -305,7 +317,9 @@ class Tracer extends NoopProxy {
       this._modules.aiguard.disable()
       this._modules.iast.disable()
       this._modules.llmobs.disable()
-      this._modules.openfeature.disable()
+      for (const feature of Tracer._featureRegistry) {
+        this._modules[feature.name].disable()
+      }
     }
 
     if (this._tracingInitialized) {
