+++ scripts/electron-package-excludes.json
@@ -4,6 +4,7 @@
     "@datadog/native-appsec",
     "@datadog/native-iast-taint-tracking",
     "@datadog/native-metrics",
+    "@datadog/openfeature-node-server",
     "@datadog/pprof",
     "@datadog/wasm-js-rewriter"
   ]
